// src/app/models/proveedores.model.ts
export interface Proveedor {
  ID_PROVEEDOR: string;
  NOMBRE_PROVEEDOR: string;
  DESCRIPCION: string;
  TELEFONO: string;
  CORREO: string;
}

/**
  Interfaz para  crear la estructura a la hora de obtener los datos.
  deben ir en mayuscula ya que cuando se realiza la consulta,
  las claves vienen dadas en mayusculas.
**/