// src/app/models/producto.modificar.model.ts
export interface ProductosModificar {
  COD_PRODUCTO: string;
  NOMBRE_PRODUCTO: string;
  COD_CATEGORIA: string;
  COD_SUB_CATEGORIA: string;
  COSTO_TOTAL_PRODUCTO: number;
  FECHA_VENCIMIENTO: string;
  ID_PROVEEDOR: string;
  DESCRIPCION: string;
  STOCK: number;
  ENUSO: number;
  ESTADO_CRITICO: number;
  NOMBRE_PRESENTACION: string;
  NOMBRE_UNIDAD: string;
  UNIDADES: number;
  MONEDA: string;
}
// Este modelo es para cuando obtengo los datos del producto que se eligen para modificar